+++ src/utils.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 import * as yaml from 'js-yaml';
 import * as fs from 'fs';
-import { sep as pathSeperator } from 'path';
+import * as path from 'path';
 
 /**
  * Checks whether the position in the line is in between curly brackets.
@@ -14,7 +14,7 @@ export function isInsideBrackets(currentLine: string, position: number): boolean
 /**
  * Checks whether string before cursor contains'{{' or not
  */
-function isBracketsInPrefix(prefix: string) {
+function isBracketsInPrefix(prefix: string): boolean {
     let prevChar = '';
     for (let index = prefix.length - 1; index >= 0; index--) {
         if (prefix.charAt(index) === '}') {
@@ -40,77 +40,138 @@ export function getWordAt(str: string, pos: number): string {
 /**
  * Retrieves the values from the `values.yaml`.
  */
-export function getValuesFromFile(document: vscode.TextDocument): any {
-    const filenames = getValueFileNamesFromConfig();
+export function getValuesFromFile(fileName: string): any {
+    const filenames = getValueFileNamesFromConfig().reverse();
+    const chartBasePath = getChartBasePath(fileName);
+    if (chartBasePath === undefined) {
+        return undefined;
+    }
+
+    let allYamlContent = '';
     for (const filename of filenames) {
-        const pathToValuesFile = getChartBasePath(document) + pathSeperator + filename;
-        if(fs.existsSync(pathToValuesFile)){
-            try {
-                const doc = yaml.safeLoad(fs.readFileSync(pathToValuesFile, 'utf8'));
-                return doc;
-              } catch (e) {
-                vscode.window.showErrorMessage('Error in \'' + filename + '\':\n' + e.message);
-                return undefined;
-            }
+        const pathToValuesFile = path.join(chartBasePath, filename);
+        if (!fs.existsSync(pathToValuesFile)) {
+            continue;
+        }
+
+        try {
+            allYamlContent += fs.readFileSync(pathToValuesFile, 'utf8') + '\n';
+        } catch (e) {
+            vscode.window.showErrorMessage('Error in \'' + filename + '\':\n' + e.message);
         }
     }
-    vscode.window.showErrorMessage('Could not locate any values.yaml. Is your values file named differently? Configure correct file name in your settings using \'helm-intellisense.customValueFileNames\'');
-    return undefined;
+    if (allYamlContent === '') {
+        vscode.window.showErrorMessage('Could not locate any values.yaml. Is your values file named differently? Configure correct file name in your settings using \'helm-intellisense.customValueFileNames\'');
+        return undefined;
+    }
+    allYamlContent = allYamlContent.replace(/---/gi, '').replace(/\.\.\./gi, '');
+    const loadedYaml = yaml.safeLoad(allYamlContent, {json: true});
+    if (typeof loadedYaml === undefined || loadedYaml === '') {
+        vscode.window.showErrorMessage('Something went wrong parsing value files');
+        return undefined;
+    }
+    return loadedYaml;
 }
 
 /**
- * Retrieves the namend-teamplate names from the `_helpers.tpl`.
+ * Retrieves the namend-teamplate names from all `*.tpl`.
  */
-export function getNamedTemplatesFromFile(document: vscode.TextDocument): string {
-    const pathToHelpersFile = getChartBasePath(document) + pathSeperator + "templates" + pathSeperator + "_helpers.tpl";
-    if(fs.existsSync(pathToHelpersFile)){
+export function getAllNamedTemplatesFromFiles(filePath: string): string[] {
+    const startPath = getChartBasePath(filePath) + path.sep + 'templates';
+    const tplFiles: string[] = getAllTplFilesFromDirectoryRecursively(startPath);
+
+    let content = '';
+    for (const tplFile of tplFiles) {
+        if (!fs.existsSync(tplFile)) {
+            continue;
+        }
         try {
-            const content: string = fs.readFileSync(pathToHelpersFile, 'utf8');
-            return content;
-            } catch (e) {
-            vscode.window.showErrorMessage('Error in \'' + pathToHelpersFile + '\':\n' + e.message);
-            return "";
+            content += fs.readFileSync(tplFile, 'utf8') + '\n\n';
+        } catch (e) {
+            vscode.window.showErrorMessage('Error in \'' + tplFile + '\':\n' + e.message);
+        }
+    }
+    return getListOfNamedTemplates(content);
+}
+
+function getAllTplFilesFromDirectoryRecursively(startPath: string): string[] {
+    if (!fs.existsSync(startPath)) {
+        return [];
+    }
+
+    let tplFiles: string[] = [];
+    const files: string[] = fs.readdirSync(startPath);
+    for (const file of files) {
+        const filename = path.join(startPath, file);
+        if (fs.lstatSync(filename).isDirectory()) {
+            tplFiles = tplFiles.concat(getAllTplFilesFromDirectoryRecursively(filename));
+        } else if (filename.indexOf('.tpl') >= 0) {
+            tplFiles.push(filename);
+        }
+    }
+    return tplFiles;
+}
+
+/**
+ * Parses named-template names from the _helpers.tpl files content.
+ */
+function getListOfNamedTemplates(content: string): string[] {
+    const matchRanges = [];
+
+    const templatePattern = /{{-? *define +"(.+?)" *-?}}/g;
+    let result;
+    while ((result = templatePattern.exec(content)) !== null) {
+        matchRanges.push(result[1]);
+    }
+    return matchRanges;
+}
+
+export function getChartBasePath(fileName: string): string | undefined {
+    if (!fs.statSync(fileName).isFile()) {
+        return undefined;
+    }
+
+    let possiblePathToChartDirectory = path.dirname(fileName);
+    const dirs = ['templates', 'charts', 'crds'];
+    for (const dir of dirs) {
+        const lastIndexOf = possiblePathToChartDirectory.lastIndexOf(path.sep + dir);
+        if (lastIndexOf !== -1) {
+            possiblePathToChartDirectory = possiblePathToChartDirectory.substr(0, lastIndexOf);
+            break;
         }
     }
 
-    vscode.window.showErrorMessage('Could not locate any _helpers.tpl.');
-    return "";
-}
-
-export function getChartBasePath(document: vscode.TextDocument): string | undefined {
-    const pathToChartDirectory = document.fileName.substr(0, document.fileName.lastIndexOf(pathSeperator + 'templates'));
-
-    if(pathToChartDirectory !== '' && fs.existsSync(pathToChartDirectory)){
-        return pathToChartDirectory;
+    const currentFilesInDir = fs.readdirSync(possiblePathToChartDirectory);
+    if (!currentFilesInDir.includes('Chart.yaml')) {
+        return undefined;
     }
 
-    const currentFilePath = document.fileName.substring(0, document.fileName.lastIndexOf(pathSeperator));
-    const currentFilesInDir = fs.readdirSync(currentFilePath);
-    if (currentFilesInDir.includes("Chart.yaml") && currentFilesInDir.includes("templates")) {
-        return currentFilePath;
-    } 
-
-    return undefined;
+    return possiblePathToChartDirectory;
 }
 
-
-export function getChartName(basePath: string): string {
-    const pathToChartFile = basePath + pathSeperator + 'Chart.yaml';
-    if(fs.existsSync(pathToChartFile)){
-        const chartYaml = yaml.safeLoad(fs.readFileSync(pathToChartFile, 'utf8'));
-        return String((chartYaml as any).name);
+export function getNameOfChart(filePath: string): string {
+    const chartBasePath = getChartBasePath(filePath);
+    if (chartBasePath === undefined) {
+        return 'No name found';
     }
-    return 'No name found';
+
+    const pathToChartFile = path.join(chartBasePath, 'Chart.yaml');
+    if (!fs.existsSync(pathToChartFile)) {
+        return 'No name found';
+    }
+
+    const chartYaml = yaml.safeLoad(fs.readFileSync(pathToChartFile, 'utf8'));
+    return String((chartYaml as any).name);
 }
 
 /**
  * Pulls list of possible values filenames from config.
  */
-function getValueFileNamesFromConfig():Array<string> {
-    const customValueFileNames:any = vscode.workspace.getConfiguration('helm-intellisense').get('customValueFileNames');
-    let filenames = [];
+function getValueFileNamesFromConfig(): string[] {
+    const customValueFileNames: any = vscode.workspace.getConfiguration('helm-intellisense').get('customValueFileNames');
+    const filenames = [];
     for (const filename of customValueFileNames) {
         filenames.push(filename);
     }
-    return filenames;	
+    return filenames;
 }